Remove mapbox-specific code
