+++ src/css/mapbox-gl.css
@@ -536,12 +536,6 @@ a.mapboxgl-ctrl-logo.mapboxgl-compact {
     text-decoration: underline;
 }
 
-/* stylelint-disable-next-line selector-class-pattern */
-.mapboxgl-ctrl-attrib .mapbox-improve-map {
-    font-weight: bold;
-    margin-left: 2px;
-}
-
 .mapboxgl-attrib-empty {
     display: none;
 }
@@ -802,11 +796,4 @@ a.mapboxgl-ctrl-logo.mapboxgl-compact {
     background: #fff;
     border: 2px dotted #202020;
     opacity: 0.5;
-}
-
-@media print {
-    /* stylelint-disable-next-line selector-class-pattern */
-    .mapbox-improve-map {
-        display: none;
-    }
 }
