+++ src/source/load_tilejson.js
@@ -5,7 +5,7 @@ import {pick, extend} from '../util/util';
 import {getJSON, ResourceType} from '../util/ajax';
 import browser from '../util/browser';
 
-import type {RequestManager} from '../util/mapbox';
+import type {RequestManager} from '../util/request_manager';
 import type {Callback} from '../types/callback';
 import type {TileJSON} from '../types/tilejson';
 import type {Cancelable} from '../types/cancelable';
@@ -26,13 +26,12 @@ export default function(options: any, requestManager: RequestManager, callback: 
                 result.vectorLayerIds = result.vectorLayers.map((layer) => { return layer.id; });
             }
 
-            result.tiles = requestManager.canonicalizeTileset(result, options.url);
             callback(null, result);
         }
     };
 
     if (options.url) {
-        return getJSON(requestManager.transformRequest(requestManager.normalizeSourceURL(options.url), ResourceType.Source), loaded);
+        return getJSON(requestManager.transformRequest(options.url, ResourceType.Source), loaded);
     } else {
         return browser.frame(() => loaded(null, options));
     }
