+++ src/style/load_glyph_range.js
@@ -5,7 +5,7 @@ import {getArrayBuffer, ResourceType} from '../util/ajax';
 import parseGlyphPBF from './parse_glyph_pbf';
 
 import type {StyleGlyph} from './style_glyph';
-import type {RequestManager} from '../util/mapbox';
+import type {RequestManager} from '../util/request_manager';
 import type {Callback} from '../types/callback';
 
 export default function (fontstack: string,
@@ -17,7 +17,7 @@ export default function (fontstack: string,
     const end = begin + 255;
 
     const request = requestManager.transformRequest(
-        requestManager.normalizeGlyphsURL(urlTemplate)
+        urlTemplate
             .replace('{fontstack}', fontstack)
             .replace('{range}', `${begin}-${end}`),
         ResourceType.Glyphs);
